requirements.txt
Pillow==9.1.0
Pillow==9.5.0
streamlit==1.11.0
protobuf==3.20.*
altair<5